+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Unreleased
 
+# 1.11.0
+
 - Fixed issue resolving dependencies if functions were passed non-string
   arguments, or returned ints.
 - Add `{{map}}` and `{{arr}}` utility template functions.
+- Fix postgres sources not working.
 
 # 1.10.0